feat:配置 localmaven 开关

+++ buildSrc/src/main/kotlin/plugin/localmaven/JarFlatLocalMaven.kt
@@ -1,6 +1,7 @@
 package plugin.localmaven
 
 import org.gradle.api.Project
+import plugin.RocketXPlugin
 import plugin.utils.FileUtil
 import java.io.File
 
@@ -12,42 +13,54 @@ import java.io.File
  */
 class JarFlatLocalMaven(
     var childProject: Project,
-    var appProject: Project,
+    var rocketXPlugin: RocketXPlugin,
     var allChangedProject: MutableMap<String, Project>? = null) : LocalMaven() {
 
     companion object {
         const val JAR = "jar"
     }
 
+    val enableLocalMaven by lazy {
+        rocketXPlugin.mRocketXBean?.localMaven ?:false
+    }
+
     override fun uploadLocalMaven() {
-        var localMavenTask = childProject.task("uploadLocalMaven") {
-            it.doLast {
-                var inputPath = FileUtil.findFirstLevelJarPath(childProject)
-                var outputFile = File(FileUtil.getLocalMavenCacheDir(), childProject.name + ".jar")
-
-                inputPath?.let {
-                    if (outputFile.exists()) {
-                        outputFile.delete()
+        if(enableLocalMaven) {
+            // publish local maven
+            try {
+                val publishTask = childProject.project.tasks.named("publishMavenJavaPublicationToLocalRepository").orNull
+                childProject.tasks.findByPath(JAR)?.let { task ->
+                    if (publishTask != null) {
+                        task.finalizedBy(publishTask)
                     }
-                    File(it).copyTo(outputFile, true)
-                    putIntoLocalMaven(childProject.name,childProject.name + ".jar")
                 }
+            } catch (e: Throwable) {
+                e.printStackTrace()
+            }
+        } else {
+            //通过 flat copy 到cache 目录
+            var localMavenTask = childProject.task("uploadLocalMaven") {
+                it.doLast {
+                    var inputPath = FileUtil.findFirstLevelJarPath(childProject)
+                    var outputFile = File(FileUtil.getLocalMavenCacheDir(), childProject.name + ".jar")
+
+                    inputPath?.let {
+                        if (outputFile.exists()) {
+                            outputFile.delete()
+                        }
+                        File(it).copyTo(outputFile, true)
+                        putIntoLocalMaven(childProject.name,childProject.name + ".jar")
+                    }
+                }
+            }
+
+            childProject.tasks.findByPath(JAR)?.let { task ->
+                 task.finalizedBy(localMavenTask)
             }
         }
 
-        // publish local maven
-        try {
-            val publishTask = childProject.project.tasks.named("publishMavenJavaPublicationToLocalRepository").orNull
-            childProject.tasks.findByPath(JAR)?.let { task ->
-                if (publishTask != null) {
-                    task.finalizedBy(localMavenTask, publishTask)
-                } else {
-                    task.finalizedBy(localMavenTask)
-                }
-            }
-        } catch (e: Throwable) {
-            e.printStackTrace()
-        }
+
+
     }
